+++ R/fs_path.R
@@ -112,6 +112,14 @@ colourise_fs_path <- function(x, ..., colors = Sys.getenv("LS_COLORS", gnu_ls_de
 
   vals <- strsplit(colors, ":")[[1]]
   nms <- strsplit(vals, "=")
+
+  if (!(
+      length(vals) == length(nms) &&
+      all(lengths(nms) == 2)
+      )) {
+    return(x)
+  }
+
   map <- setNames(
     vapply(nms, `[[`, character(1), 2),
     vapply(nms, `[[`, character(1), 1))